requirements.txt
streamlit==1.29.0
supabase==2.0.3
plotly==5.17.0
pandas==2.3.2
python-dotenv==1.0.0